python-backend/pyproject.toml
[build-system]
requires = ["flit_core >=3.2", "packaging~=23.1", "setuptools~=67.8", "wheel~=0.40"]
build-backend = "flit_core.buildapi"

[project]
name = "videoseal"
requires-python = ">=3.9,<3.12"
dynamic = ["version", "description"]

dependencies = [
    "setuptools",
    "torch>=2.3.1",
    "pandas",
    "opencv-python",
    "omegaconf",
    "einops",
    "lpips",
    "timm==0.9.16",
    "pre-commit",
    "ipykernel",
    "pycocotools",
    "pandas",
    "PyWavelets",
    "av",
    "pyav",
    "scikit-image",
    "pandas",
    "tqdm",
    "pytorch_msssim",
    "tensorboard",
    "calflops",
    "transformers",
    "safetensors",
    "pycocotools",
    "opencv-python",
    "tqdm",
    "timm",
    "scipy",
    "ffmpeg-python",
    "torchcodec>=0.1.1",
    "torchaudio>=2.5.1",
    "audioseal>=0.1.4",
    "flask>=3.1.0",
    "requests-toolbelt>=1.0.0",
]

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cu121", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu121", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchcodec = [
  { index = "pytorch-cu121", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchaudio = [
  { index = "pytorch-cu121", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[project.optional-dependencies]
  dev = [
    "func_argparse",
    "pytest",
    "black",
    "isort",
    "flake8",
    "pre-commit",
    "conda-pack",
  ]

[tool.flake8]
extend_ignore = ["E", "Y"]  # Black
per-file-ignores = [
    "__init__.py:F401",
]

[tool.isort]
profile = "black"
